refactor: extrai camada de serviço nos 4 pacotes CLI (RFC 0013 Fase 2)
Cada pacote (djen_backup, tjro_juris, stj_acordaos, datajud) ganha um
service.py com funções config→resultado, sem Typer/Rich/echo — os
__main__.py ficam reduzidos a parsing de argv e tradução do resultado em
echo/exit code.

- djen_backup: side effects de import (structlog.configure, reconfiguração
  de stdout/stderr) saem do topo do módulo para configure_runtime(),
  chamada uma vez no callback main() — Click invoca o callback do grupo
  antes de qualquer subcomando, cobrindo check/upload/drain/probe/reset
  também. O bug de exit code descartado — documentado só no callback bare —
  também existia em check e upload; os três agora fazem
  raise typer.Exit(code=_run_pipeline(...)).
- tjro_juris: crawl/upload/status/consolidate movidos quase inteiros para
  service.py. crawl_bounds passa a levantar ValueError (framework-neutro)
  em vez de typer.BadParameter; o __main__.py traduz na borda da CLI.
- stj_acordaos: download_one/upload_all retornam outcomes estruturados em
  vez de chamar typer.echo diretamente. ia_key/ia_secret deixam de ser
  opção de CLI — upload só lê IA_ACCESS_KEY/IA_SECRET_KEY do ambiente via
  service.ia_credentials().
- datajud: mesmo padrão para enrich/facetas/status; ia_key/ia_secret
  removidos da opção enrich pelo mesmo motivo.

Testes de caracterização da Fase 1 continuam verdes, exceto os dois que
documentavam ia_key/ia_secret como opção de CLI — atualizados para afirmar
a ausência da opção, já que essa mudança de contrato é o objetivo desta
fase. Testes unitários que monkeypatchavam símbolos de __main__.py foram
atualizados para os novos módulos de destino.

751 testes, 0 falhas, 0 erros, 1 skipped. ruff check/format limpos.

Não incluído: a bateria de testes framework-neutra (argv → configuração
semântica) cotada como portão durável de Fase 4 — fica para o início da
Fase 4, quando a API real do Cyclopts for conhecida.
