+++ ruff.toml
@@ -60,7 +60,21 @@ ignore = [
 # E501: long IA metadata strings; re-wrapping would harm readability.
 # BLE001: blind except around a single resource's download/extract in a batch
 # CLI loop — one bad ZIP shouldn't abort the whole run.
-"src/stj_acordaos/__main__.py" = ["B008", "PLC0415", "D101", "D102", "D103", "D107", "FBT001", "FBT003", "BLE001"]
+# TC003: Path used at runtime — Typer resolves bare `param: Path = ...`
+# annotations via get_type_hints(), so the import can't move into
+# TYPE_CHECKING (same rationale as tjro_juris/__main__.py below).
+"src/stj_acordaos/__main__.py" = [
+    "B008",
+    "PLC0415",
+    "D101",
+    "D102",
+    "D103",
+    "D107",
+    "FBT001",
+    "FBT003",
+    "BLE001",
+    "TC003",
+]
 "src/stj_acordaos/manifest.py" = ["D107", "D105", "PLW2901", "PERF401"]
 # S608: f-string SQL with DuckDB — path_list is internal (no user input).
 # E501: long log line; splitting would obscure the structured-logging pattern.
@@ -82,11 +96,15 @@ ignore = [
 "src/tjro_juris/archive.py" = ["D100", "D103", "PLC0415", "ASYNC240", "TC003"]
 
 # ── datajud ───────────────────────────────────────────────────────────────
-# PLC0415: lazy duckdb imports inside CLI helpers — standard pattern.
 # FBT001/FBT002: boolean CLI flags are idiomatic Typer.
+# TC003: Path used at runtime — Typer resolves `Annotated[Path, ...]`
+# annotations via get_type_hints(), so the import can't move into
+# TYPE_CHECKING (same rationale as tjro_juris/__main__.py above).
+"src/datajud/__main__.py" = ["FBT001", "FBT002", "TC003"]
+# PLC0415: lazy duckdb imports inside data-gathering helpers — standard pattern.
 # S608: f-string SQL over internal parquet paths only (no user input).
-# S310: urlopen against a fixed archive.org URL constant, not user input.
-"src/datajud/__main__.py" = ["PLC0415", "FBT001", "FBT002", "S608", "S310"]
+# S310: urlopen against a fixed archive.org URL constant, not user input (inline noqa).
+"src/datajud/service.py" = ["PLC0415", "S608"]
 
 # ── Lazy / conditional imports in library modules ─────────────────────────
 # These modules defer expensive imports to avoid slowing every process start.
